도연/12.19/Api수정

+++ src/main/java/com/mute/Final/service/DeleteService.java
@@ -21,15 +21,24 @@ public class DeleteService {
     private final ReviewSeatRepository reviewSeatRepository; // 좌석 후기
 
     // 총평 후기 삭제
-    public boolean deleteTotal(String reviewMuId) {
+    public boolean deleteTotal(String member,String reviewMuId) {
         try {
-            Long result = reviewTotalRepository.deleteByReviewMuId(Long.parseLong(reviewMuId));
-            if (result==1) return true;
-            else return false;
+            System.out.println("here ok asdfasdf");
+            ReviewTotal reviewTotal =  reviewTotalRepository.findByReviewMuId(Long.parseLong(reviewMuId));
+            System.out.println("findByReviewMuid 12341234 : " + reviewTotal);
+
+            if(reviewTotal.getMember().getUserNum().equals(Long.parseLong(member))){
+                Long result = reviewTotalRepository.deleteByReviewMuId(Long.parseLong(reviewMuId));
+                if (result==1) return true;
+                else return false;
+            }
+
+
         } catch (Exception e) {
             log.warn("실패!!!!!!!!!!!!!ㅠㅠㅠㅠㅠㅠ");
             return false;
         }
+        return false;
     }
 
     // 좌석 후기 삭제
